+++ src/main/java/com/github/halab4dev/mongo2j/annotation/BsonIgnore.java
@@ -1,0 +1,14 @@
+package com.github.halab4dev.mongo2j.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * author halab
+ */
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface BsonIgnore {
+}